+++ release_notes.md
@@ -1,4 +1,9 @@
 ## Changes
 
+### 0.0.0-alpha.2
+- Updated `aws-stream-consumer-core` dependency to version 2.0.5
+- Minor fixes to type definitions & JsDoc comments
+- Renamed `processStreamEvent` parameters to match those of `kinesis-stream-consumer`
+
 ### 0.0.0-alpha.1
 - Refactored and extracted DynamoDB stream consumer-specific logic from `aws-stream-consumer` into this new module